+++ tests/test_preferencehandler.py
@@ -36,6 +36,7 @@ from ospd_openvas.preferencehandler import (
     BOREAS_ALIVE_TEST,
     BOREAS_ALIVE_TEST_PORTS,
     PreferenceHandler,
+    alive_test_methods_to_bit_field,
 )
 
 
@@ -203,6 +204,13 @@ class PreferenceHandlerTestCase(TestCase):
 
         self.assertEqual(ret, [])
 
+        # alive test was supplied via seperate xml element
+        w = DummyDaemon()
+        target_options_dict = {'alive_test_methods': '1', 'icmp': '0'}
+        p = PreferenceHandler('1234-1234', None, w.scan_collection, None)
+        ret = p.build_alive_test_opt_as_prefs(target_options_dict)
+        self.assertEqual(ret, [])
+
     def test_build_alive_test_opt(self):
         w = DummyDaemon()
 
@@ -218,6 +226,21 @@ class PreferenceHandlerTestCase(TestCase):
         p = PreferenceHandler('1234-1234', None, w.scan_collection, None)
         ret = p.build_alive_test_opt_as_prefs(target_options_dict)
 
+        self.assertEqual(ret, alive_test_out)
+
+        # alive test was supplied via sepertae xml element
+        w = DummyDaemon()
+        alive_test_out = [
+            "1.3.6.1.4.1.25623.1.0.100315:1:checkbox:Do a TCP ping|||no",
+            "1.3.6.1.4.1.25623.1.0.100315:2:checkbox:TCP ping tries also TCP-SYN ping|||no",
+            "1.3.6.1.4.1.25623.1.0.100315:7:checkbox:TCP ping tries only TCP-SYN ping|||no",
+            "1.3.6.1.4.1.25623.1.0.100315:3:checkbox:Do an ICMP ping|||yes",
+            "1.3.6.1.4.1.25623.1.0.100315:4:checkbox:Use ARP|||no",
+            "1.3.6.1.4.1.25623.1.0.100315:5:checkbox:Mark unrechable Hosts as dead (not scanning)|||yes",
+        ]
+        target_options_dict = {'alive_test_methods': '1', 'icmp': '1'}
+        p = PreferenceHandler('1234-1234', None, w.scan_collection, None)
+        ret = p.build_alive_test_opt_as_prefs(target_options_dict)
         self.assertEqual(ret, alive_test_out)
 
     def test_build_alive_test_opt_fail_1(self):
@@ -243,7 +266,8 @@ class PreferenceHandlerTestCase(TestCase):
         p.prepare_target_for_openvas()
 
         p.kbdb.add_scan_preferences.assert_called_with(
-            p.scan_id, ['TARGET|||192.168.0.1'],
+            p.scan_id,
+            ['TARGET|||192.168.0.1'],
         )
 
     @patch('ospd_openvas.db.KbDB')
@@ -258,7 +282,8 @@ class PreferenceHandlerTestCase(TestCase):
         p.prepare_ports_for_openvas()
 
         p.kbdb.add_scan_preferences.assert_called_with(
-            p.scan_id, ['port_range|||80,443'],
+            p.scan_id,
+            ['port_range|||80,443'],
         )
 
     @patch('ospd_openvas.db.KbDB')
@@ -271,7 +296,8 @@ class PreferenceHandlerTestCase(TestCase):
         p.prepare_main_kbindex_for_openvas()
 
         p.kbdb.add_scan_preferences.assert_called_with(
-            p.scan_id, ['ov_maindbid|||2'],
+            p.scan_id,
+            ['ov_maindbid|||2'],
         )
 
     @patch('ospd_openvas.db.KbDB')
@@ -364,7 +390,8 @@ class PreferenceHandlerTestCase(TestCase):
         p.prepare_host_options_for_openvas()
 
         p.kbdb.add_scan_preferences.assert_called_with(
-            p.scan_id, ['exclude_hosts|||192.168.0.1'],
+            p.scan_id,
+            ['exclude_hosts|||192.168.0.1'],
         )
 
     @patch('ospd_openvas.db.KbDB')
@@ -423,7 +450,10 @@ class PreferenceHandlerTestCase(TestCase):
 
         p.kbdb.add_scan_preferences.assert_called_with(
             p.scan_id,
-            ['reverse_lookup_only|||yes', 'reverse_lookup_unify|||no',],
+            [
+                'reverse_lookup_only|||yes',
+                'reverse_lookup_unify|||no',
+            ],
         )
 
     @patch('ospd_openvas.db.KbDB')
@@ -431,6 +461,8 @@ class PreferenceHandlerTestCase(TestCase):
         # No Boreas config setting (BOREAS_SETTING_NAME) set
         w = DummyDaemon()
         ov_setting = {'not_the_correct_setting': 1}
+        t_opt = {}
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
         with patch.object(Openvas, 'get_settings', return_value=ov_setting):
             p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
             p.scan_id = '456-789'
@@ -450,7 +482,8 @@ class PreferenceHandlerTestCase(TestCase):
             p.kbdb.add_scan_preferences = MagicMock()
             p.prepare_boreas_alive_test()
 
-            p.kbdb.add_scan_preferences.assert_not_called()
+            calls = [call(p.scan_id, [BOREAS_ALIVE_TEST + '|||2'])]
+            p.kbdb.add_scan_preferences.assert_has_calls(calls)
 
         # ALIVE_TEST_TCP_SYN_SERVICE as alive test.
         w = DummyDaemon()
@@ -515,6 +548,167 @@ class PreferenceHandlerTestCase(TestCase):
             p.kbdb.add_scan_preferences.assert_has_calls(calls)
 
     @patch('ospd_openvas.db.KbDB')
+    def test_set_boreas_alive_test_not_as_enum(self, mock_kb):
+        # No Boreas config setting (BOREAS_SETTING_NAME) set
+        w = DummyDaemon()
+        ov_setting = {'not_the_correct_setting': 1}
+        t_opt = {}
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
+        with patch.object(Openvas, 'get_settings', return_value=ov_setting):
+            p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
+            p.scan_id = '456-789'
+            p.kbdb.add_scan_preferences = MagicMock()
+            p.prepare_boreas_alive_test()
+
+            p.kbdb.add_scan_preferences.assert_not_called()
+
+        # Boreas config setting set but invalid alive_test.
+        w = DummyDaemon()
+        t_opt = {'alive_test_methods': "1", 'arp': '-1'}
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
+        ov_setting = {BOREAS_SETTING_NAME: 1}
+        with patch.object(Openvas, 'get_settings', return_value=ov_setting):
+            p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
+            p.scan_id = '456-789'
+            p.kbdb.add_scan_preferences = MagicMock()
+            p.prepare_boreas_alive_test()
+
+            calls = [call(p.scan_id, [BOREAS_ALIVE_TEST + '|||2'])]
+            p.kbdb.add_scan_preferences.assert_has_calls(calls)
+
+        # ICMP was chosen as alive test.
+        w = DummyDaemon()
+        t_opt = {'alive_test_methods': "1", 'icmp': '1'}
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
+        ov_setting = {BOREAS_SETTING_NAME: 1}
+        with patch.object(Openvas, 'get_settings', return_value=ov_setting):
+            p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
+            p.scan_id = '456-789'
+            p.kbdb.add_scan_preferences = MagicMock()
+            p.prepare_boreas_alive_test()
+
+            calls = [call(p.scan_id, [BOREAS_ALIVE_TEST + '|||2'])]
+            p.kbdb.add_scan_preferences.assert_has_calls(calls)
+
+        # tcp_syn as alive test.
+        w = DummyDaemon()
+        t_opt = {'alive_test_methods': "1", 'tcp_syn': '1'}
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
+        ov_setting = {BOREAS_SETTING_NAME: 1}
+        with patch.object(Openvas, 'get_settings', return_value=ov_setting):
+            p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
+            p.scan_id = '456-789'
+            p.kbdb.add_scan_preferences = MagicMock()
+            p.prepare_boreas_alive_test()
+
+            calls = [call(p.scan_id, [BOREAS_ALIVE_TEST + '|||16'])]
+            p.kbdb.add_scan_preferences.assert_has_calls(calls)
+
+        # tcp_ack as alive test.
+        w = DummyDaemon()
+        t_opt = {'alive_test_methods': "1", 'tcp_ack': '1'}
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
+        ov_setting = {BOREAS_SETTING_NAME: 1}
+        with patch.object(Openvas, 'get_settings', return_value=ov_setting):
+            p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
+            p.scan_id = '456-789'
+            p.kbdb.add_scan_preferences = MagicMock()
+            p.prepare_boreas_alive_test()
+
+            calls = [call(p.scan_id, [BOREAS_ALIVE_TEST + '|||1'])]
+            p.kbdb.add_scan_preferences.assert_has_calls(calls)
+
+        # arp as alive test.
+        w = DummyDaemon()
+        t_opt = {'alive_test_methods': "1", 'arp': '1'}
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
+        ov_setting = {BOREAS_SETTING_NAME: 1}
+        with patch.object(Openvas, 'get_settings', return_value=ov_setting):
+            p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
+            p.scan_id = '456-789'
+            p.kbdb.add_scan_preferences = MagicMock()
+            p.prepare_boreas_alive_test()
+
+            calls = [call(p.scan_id, [BOREAS_ALIVE_TEST + '|||4'])]
+            p.kbdb.add_scan_preferences.assert_has_calls(calls)
+
+        # arp as alive test.
+        w = DummyDaemon()
+        t_opt = {'alive_test_methods': "1", 'consider_alive': '1'}
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
+        ov_setting = {BOREAS_SETTING_NAME: 1}
+        with patch.object(Openvas, 'get_settings', return_value=ov_setting):
+            p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
+            p.scan_id = '456-789'
+            p.kbdb.add_scan_preferences = MagicMock()
+            p.prepare_boreas_alive_test()
+
+            calls = [call(p.scan_id, [BOREAS_ALIVE_TEST + '|||8'])]
+            p.kbdb.add_scan_preferences.assert_has_calls(calls)
+
+        # all alive test methods
+        w = DummyDaemon()
+        t_opt = {
+            'alive_test_methods': "1",
+            'icmp': '1',
+            'tcp_ack': '1',
+            'tcp_syn': '1',
+            'arp': '1',
+            'consider_alive': '1',
+        }
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
+        ov_setting = {BOREAS_SETTING_NAME: 1}
+        with patch.object(Openvas, 'get_settings', return_value=ov_setting):
+            p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
+            p.scan_id = '456-789'
+            p.kbdb.add_scan_preferences = MagicMock()
+            p.prepare_boreas_alive_test()
+
+            calls = [call(p.scan_id, [BOREAS_ALIVE_TEST + '|||31'])]
+            p.kbdb.add_scan_preferences.assert_has_calls(calls)
+
+        # TCP-SYN alive test and dedicated port list for alive scan provided.
+        w = DummyDaemon()
+        t_opt = {
+            'alive_test_ports': "80,137",
+            'alive_test_methods': "1",
+            'tcp_syn': '1',
+        }
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
+        ov_setting = {BOREAS_SETTING_NAME: 1}
+        with patch.object(Openvas, 'get_settings', return_value=ov_setting):
+            p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
+            p.scan_id = '456-789'
+            p.kbdb.add_scan_preferences = MagicMock()
+            p.prepare_boreas_alive_test()
+
+            calls = [
+                call(p.scan_id, [BOREAS_ALIVE_TEST + '|||16']),
+                call(p.scan_id, [BOREAS_ALIVE_TEST_PORTS + '|||80,137']),
+            ]
+            p.kbdb.add_scan_preferences.assert_has_calls(calls)
+
+    @patch('ospd_openvas.db.KbDB')
+    def test_set_boreas_alive_test_enum_has_precedence(self, mock_kb):
+        w = DummyDaemon()
+        t_opt = {
+            'alive_test_methods': "1",
+            'consider_alive': '1',
+            'alive_test': AliveTest.ALIVE_TEST_ICMP,
+        }
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
+        ov_setting = {BOREAS_SETTING_NAME: 1}
+        with patch.object(Openvas, 'get_settings', return_value=ov_setting):
+            p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
+            p.scan_id = '456-789'
+            p.kbdb.add_scan_preferences = MagicMock()
+            p.prepare_boreas_alive_test()
+
+            # has icmp and not consider_alive
+            calls = [call(p.scan_id, [BOREAS_ALIVE_TEST + '|||2'])]
+            p.kbdb.add_scan_preferences.assert_has_calls(calls)
+
+    @patch('ospd_openvas.db.KbDB')
     def test_set_boreas_alive_test_without_settings(self, mock_kb):
         w = DummyDaemon()
         t_opt = {'alive_test': 16}
@@ -536,6 +730,40 @@ class PreferenceHandlerTestCase(TestCase):
         w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
 
         ov_setting = {}
+
+        with patch.object(Openvas, 'get_settings', return_value=ov_setting):
+            p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
+            p.scan_id = '456-789'
+            p.kbdb.add_scan_preferences = MagicMock()
+            p.prepare_alive_test_option_for_openvas()
+
+            p.kbdb.add_scan_preferences.assert_not_called()
+
+    @patch('ospd_openvas.db.KbDB')
+    def test_set_alive_no_invalid_alive_test(self, mock_kb):
+        w = DummyDaemon()
+
+        t_opt = {'alive_test': -1}
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
+
+        ov_setting = {'some_setting': 1}
+
+        with patch.object(Openvas, 'get_settings', return_value=ov_setting):
+            p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
+            p.scan_id = '456-789'
+            p.kbdb.add_scan_preferences = MagicMock()
+            p.prepare_alive_test_option_for_openvas()
+
+            p.kbdb.add_scan_preferences.assert_not_called()
+
+    @patch('ospd_openvas.db.KbDB')
+    def test_set_alive_no_invalid_alive_test_no_enum(self, mock_kb):
+        w = DummyDaemon()
+
+        t_opt = {'alive_test_methods': '1', 'icmp': '-1'}
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
+
+        ov_setting = {'some_setting': 1}
 
         with patch.object(Openvas, 'get_settings', return_value=ov_setting):
             p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
@@ -571,5 +799,130 @@ class PreferenceHandlerTestCase(TestCase):
             p.prepare_alive_test_option_for_openvas()
 
             p.kbdb.add_scan_preferences.assert_called_with(
-                p.scan_id, alive_test_out,
+                p.scan_id,
+                alive_test_out,
             )
+
+    @patch('ospd_openvas.db.KbDB')
+    def test_prepare_alive_test_not_supplied_as_enum(self, mock_kb):
+        w = DummyDaemon()
+
+        alive_test_out = [
+            "1.3.6.1.4.1.25623.1.0.100315:1:checkbox:Do a TCP ping|||no",
+            "1.3.6.1.4.1.25623.1.0.100315:2:checkbox:TCP ping tries also TCP-SYN ping|||no",
+            "1.3.6.1.4.1.25623.1.0.100315:7:checkbox:TCP ping tries only TCP-SYN ping|||no",
+            "1.3.6.1.4.1.25623.1.0.100315:3:checkbox:Do an ICMP ping|||yes",
+            "1.3.6.1.4.1.25623.1.0.100315:4:checkbox:Use ARP|||no",
+            "1.3.6.1.4.1.25623.1.0.100315:5:checkbox:Mark unrechable Hosts as dead (not scanning)|||yes",
+        ]
+
+        t_opt = {'alive_test_methods': '1', 'icmp': '1'}
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
+
+        ov_setting = {'some_setting': 1}
+
+        with patch.object(Openvas, 'get_settings', return_value=ov_setting):
+            p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
+
+            p.scan_id = '456-789'
+            p.kbdb.add_scan_preferences = MagicMock()
+            p.prepare_alive_test_option_for_openvas()
+
+            p.kbdb.add_scan_preferences.assert_called_with(
+                p.scan_id,
+                alive_test_out,
+            )
+
+    @patch('ospd_openvas.db.KbDB')
+    def test_prepare_alive_test_no_enum_no_alive_test(self, mock_kb):
+        w = DummyDaemon()
+
+        t_opt = {'alive_test_methods': '1', 'icmp': '0'}
+        w.scan_collection.get_target_options = MagicMock(return_value=t_opt)
+
+        ov_setting = {'some_setting': 1}
+
+        with patch.object(Openvas, 'get_settings', return_value=ov_setting):
+            p = PreferenceHandler('1234-1234', mock_kb, w.scan_collection, None)
+
+            p.scan_id = '456-789'
+            p.kbdb.add_scan_preferences = MagicMock()
+            p.prepare_alive_test_option_for_openvas()
+
+            p.kbdb.add_scan_preferences.assert_not_called()
+
+    def test_alive_test_methods_to_bit_field(self):
+
+        self.assertEqual(
+            AliveTest.ALIVE_TEST_TCP_ACK_SERVICE,
+            alive_test_methods_to_bit_field(
+                icmp=False,
+                tcp_ack=True,
+                tcp_syn=False,
+                arp=False,
+                consider_alive=False,
+            ),
+        )
+
+        self.assertEqual(
+            AliveTest.ALIVE_TEST_ICMP,
+            alive_test_methods_to_bit_field(
+                icmp=True,
+                tcp_ack=False,
+                tcp_syn=False,
+                arp=False,
+                consider_alive=False,
+            ),
+        )
+
+        self.assertEqual(
+            AliveTest.ALIVE_TEST_ARP,
+            alive_test_methods_to_bit_field(
+                icmp=False,
+                tcp_ack=False,
+                tcp_syn=False,
+                arp=True,
+                consider_alive=False,
+            ),
+        )
+
+        self.assertEqual(
+            AliveTest.ALIVE_TEST_CONSIDER_ALIVE,
+            alive_test_methods_to_bit_field(
+                icmp=False,
+                tcp_ack=False,
+                tcp_syn=False,
+                arp=False,
+                consider_alive=True,
+            ),
+        )
+
+        self.assertEqual(
+            AliveTest.ALIVE_TEST_TCP_SYN_SERVICE,
+            alive_test_methods_to_bit_field(
+                icmp=False,
+                tcp_ack=False,
+                tcp_syn=True,
+                arp=False,
+                consider_alive=False,
+            ),
+        )
+
+        all_alive_test_methods = (
+            AliveTest.ALIVE_TEST_SCAN_CONFIG_DEFAULT
+            | AliveTest.ALIVE_TEST_TCP_ACK_SERVICE
+            | AliveTest.ALIVE_TEST_ICMP
+            | AliveTest.ALIVE_TEST_ARP
+            | AliveTest.ALIVE_TEST_CONSIDER_ALIVE
+            | AliveTest.ALIVE_TEST_TCP_SYN_SERVICE
+        )
+        self.assertEqual(
+            all_alive_test_methods,
+            alive_test_methods_to_bit_field(
+                icmp=True,
+                tcp_ack=True,
+                tcp_syn=True,
+                arp=True,
+                consider_alive=True,
+            ),
+        )